pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "MBTI-in-Thoughts"
version = "1.0.0"
authors = [
  { name="Maciej Besta", email="maciej.besta@inf.ethz.ch" },
  { name="Shriram Chandran", email="schandran@student.ethz.ch" },
  { name="Robert Gerstenberger", email="gerstenberger.robert@gmail.com" },
  { name="Mathis Lindner", email="mlindner@student.ethz.ch" },
  { name="Sebastian Hermann Martschat", email="sebastian.martschat@basf.com" },
  { name="Taraneh Ghandi", email="taraneh.ghandi@partners.basf.com" }
]
description = "Python package for the MBTI-in-Thoughts framework to enhancing the effectiveness of LLM agents through psychologically grounded personality conditioning."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9,<3.14"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  "aiohappyeyeballs==2.6.1",
  "aiohttp==3.11.18",
  "aiosignal==1.3.2",
  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "asttokens==3.0.0",
  "attrs==25.3.0",
  "certifi==2025.1.31",
  "charset-normalizer==3.4.1",
  "comm==0.2.2",
  "contourpy==1.3.2",
  "cycler==0.12.1",
  "dataclasses-json==0.6.7",
  "datasets==3.6.0",
  "debugpy==1.8.14",
  "decorator==5.2.1",
  "dill==0.3.8",
  "distro==1.9.0",
  "executing==2.2.0",
  "filelock==3.18.0",
  "fonttools==4.57.0",
  "frozenlist==1.6.0",
  "fsspec==2024.12.0",
  "greenlet==3.2.0",
  "h11==0.14.0",
  "httpcore==1.0.8",
  "httpx==0.28.1",
  "httpx-sse==0.4.0",
  "huggingface-hub==0.30.2",
  "idna==3.10",
  "ipykernel==6.29.5",
  "ipython==9.1.0",
  "ipython_pygments_lexers==1.1.1",
  "jedi==0.19.2",
  "Jinja2==3.1.6",
  "jiter==0.9.0",
  "joblib==1.4.2",
  "jsonpatch==1.33",
  "jsonpointer==3.0.0",
  "jupyter_client==8.6.3",
  "jupyter_core==5.7.2",
  "kiwisolver==1.4.8",
  "langchain==0.3.26",
  "langchain-anthropic",
  "langchain-community==0.3.26",
  "langchain-core==0.3.66",
  "langchain-deepseek",
  "langchain-google-genai",
  "langchain-google-vertexai",
  "langchain-huggingface==0.3.0",
  "langchain-openai==0.3.24",
  "langchain-text-splitters==0.3.8",
  "langgraph==0.4.8",
  "langgraph-checkpoint==2.1.0",
  "langgraph-prebuilt==0.2.2",
  "langgraph-sdk==0.1.61",
  "langsmith==0.4.1",
  "lexicalrichness",
  "MarkupSafe==3.0.2",
  "marshmallow==3.26.1",
  "matplotlib==3.10.1",
  "matplotlib-inline==0.1.7",
  "mpmath==1.3.0",
  "multidict==6.4.3",
  "multiprocess==0.70.16",
  "mypy-extensions==1.0.0",
  "nest-asyncio==1.6.0",
  "networkx==3.4.2",
  "nltk",
  "numpy==2.2.5",
  "nvidia-cublas-cu12==12.4.5.8",
  "nvidia-cuda-cupti-cu12==12.4.127",
  "nvidia-cuda-nvrtc-cu12==12.4.127",
  "nvidia-cuda-runtime-cu12==12.4.127",
  "nvidia-cudnn-cu12==9.1.0.70",
  "nvidia-cufft-cu12==11.2.1.3",
  "nvidia-curand-cu12==10.3.5.147",
  "nvidia-cusolver-cu12==11.6.1.9",
  "nvidia-cusparse-cu12==12.3.1.170",
  "nvidia-cusparselt-cu12==0.6.2",
  "nvidia-nccl-cu12==2.21.5",
  "nvidia-nvjitlink-cu12==12.4.127",
  "nvidia-nvtx-cu12==12.4.127",
  "openai==1.90.0",
  "openpyxl",
  "orjson==3.10.16",
  "ormsgpack==1.10.0",
  "packaging==24.2",
  "pandas==2.2.3",
  "parso==0.8.4",
  "pexpect==4.9.0",
  "pillow==11.2.1",
  "platformdirs==4.3.7",
  "praw",
  "prompt_toolkit==3.0.51",
  "propcache==0.3.1",
  "psutil==7.0.0",
  "ptyprocess==0.7.0",
  "pure_eval==0.2.3",
  "py-readability-metrics",
  "pyarrow==19.0.1",
  "pydantic==2.11.3",
  "pydantic-settings==2.9.1",
  "pydantic_core==2.33.1",
  "Pygments==2.19.1",
  "pyparsing==3.2.3",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.1.0",
  "pytz==2025.2",
  "PyYAML==6.0.2",
  "pyzmq==26.4.0",
  "readability",
  "regex==2024.11.6",
  "requests==2.32.3",
  "requests-toolbelt==1.0.0",
  "safetensors==0.5.3",
  "scikit-learn==1.6.1",
  "scipy==1.15.2",
  "seaborn==0.13.2",
  "sentence-transformers==4.1.0",
  "setuptools==79.0.0",
  "six==1.17.0",
  "sniffio==1.3.1",
  "SQLAlchemy==2.0.40",
  "stack-data==0.6.3",
  "sympy==1.13.1",
  "tenacity==9.1.2",
  "threadpoolctl==3.6.0",
  "tiktoken==0.9.0",
  "tokenizers==0.21.1",
  "torch==2.6.0",
  "tornado==6.4.2",
  "tqdm==4.67.1",
  "traitlets==5.14.3",
  "transformers==4.51.3",
  "triton==3.2.0",
  "typing-inspect==0.9.0",
  "typing-inspection==0.4.0",
  "typing_extensions==4.13.2",
  "tzdata==2025.2",
  "urllib3==2.4.0",
  "wcwidth==0.2.13",
  "xxhash==3.5.0",
  "yarl==1.20.0",
  "zstandard==0.23.0"
]

[project.urls]
Homepage = "https://github.com/spcl/MBTI-in-Thoughts"
